+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 ### Fixed
 
 * 500 result code on tx applying cause after failing of operation next operation has invalid result
+* Exception on checkSaleState with cause allowing pre issuance excess tokens after sale creation
+* Exception on check signature when there is no signers for account
+* Exception on review pre issuance when available for issuance more than max issuance
 
 ## 3.0.0-x.8
 
@@ -26,6 +29,7 @@
 * Field with payload for txNO_ROLE_PERMISSION and opNO_ROLE_PERMISSION
 * Licensing
 * License, Stamp operations
+
 ### Fixed
 
 * Do not add to cache if load entry without delta (add delta to putCacheEntry method)